.gitlab: add directory and CI YAML workflow refactoring
Problem: the .gitlab-ci.yml top-level file will become overgrown
as we progress in the number of tests that we want to run on our
framework repositories. Adding jobs, machines, and tests should be
organized and extensible.

Utilize YAML syntax and include statements to rework the files
in a directory structure.