mpi: add basic MPI hello world testing
Problem: to test MPI implementations and compilers, we need
some kind of basic test that utilizes MPI.

Add a C source file and Makefile that create an executable
to test MPI implementations. Very general so that it will work
with many different compilers. This is based on/from the MPI
tutorial on hpc.llnl.gov.