.github: add commit validator
Problem: RFC 1 outlines specific requirements for framework
repositories. This includes directives about commit messages.
Sometimes, it's easy to forget about these requirements.

Add the PR validator workflow for this repository in GitHub, so
that every PR's commit messages will be checked. It's much easier
to comply with RFC 1 if an external tool is checking for commit
messages.