gitlab: add testing for flux-sched on system instance clusters
Problem: We need to be testing flux-sched's integration with the
rest of the framework, and we're not.

Do that on corona and tioga since they're the system-instance clusters.

Note that due to missing RPATH entries for certain libraries in Cray's
compilers, this currently needs to be compiled with GCC. Also, the
Valgrind test is temporarily exempted because it fails due to a memory
leak in hwloc.