mpi: add test that physical cores are not oversubscribed
Problem: A common failure in some MPI implementations is that
all the MPI ranks end up packed on the same core. This causes
lock contention, achieves no parallelization, and is bad, so we
want to catch if that's happening.

Add a test to our CI that collects the physical cpu number for
each MPI rank, then checks that no two are the same. Note that this
will fail if cores are deliberately oversubscribed (but they
shouldn't be in CI).

Co-authored-by: Nathan Hanford <8302958+nhanford@users.noreply.github.com>