gitlab: drop HWLOC_COMPONENTS=x86
Problem: per the hwloc(7) documentation for hwloc v2.3.0, the
HWLOC_COMPONENTS environment variable, if used, will override
the HWLOC_XMLFILE variable. According to the docs, "since this
variable is the low-level and more generic way to select components,
it takes precedence over environment variables for selecting components."

I don't believe we've actually hit this problem. But just in case,
it seems advisable to drop the HWLOC_COMPONENTS variable in favor of
using the xml file.